pyproject.toml
[tool.poetry]
name = "deepmode-backend"
version = "0.1.0"
description = ""
authors = ["AlekseyLebedik <lebedik.leha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "1.12.0"
annotated-types = "0.6.0"
anyio = "3.7.1"
asyncpg = "0.28.0"
bcrypt = "4.0.1"
click = "8.1.7"
dnspython = "2.4.2"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"
envparse = "0.2.0"
exceptiongroup = "1.1.3"
fastapi = "0.103.2"
graphql-core = "3.2.3"
greenlet = "3.0.0"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.4"
libcst = "1.1.0"
mako = "1.3.0"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
mdurl = "0.1.2"
mypy-extensions = "1.0.0"
passlib = "1.7.4"
pendulum = "3.0.0"
pillow = "10.1.0"
pyasn1 = "0.5.1"
pydantic = {version = "2.5.3", extras = ["email"]}
pydantic-core = "2.14.6"
pydantic-settings = "2.0.3"
pygments = "2.17.2"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pyyaml = "6.0.1"
rich = "13.7.0"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.22"
starlette = "0.27.0"
strawberry-graphql = {version = "0.217.1", extras = ["debug-server"]}
time-machine = "2.13.0"
typer = "0.9.0"
typing-extensions = "4.8.0"
typing-inspect = "0.9.0"
tzdata = "2023.4"
tzlocal = "5.2"
uvicorn = {version = "0.25.0", extras = ["standard"]}
uvloop = "0.18.0"
watchfiles = "0.21.0"
websockets = "11.0.3"
poetry-plugin-export = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
